backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
gunicorn==20.1.0
sqlparse==0.4.3
djangorestframework==3.14.0
psycopg2
djangorestframework_simplejwt==5.2.1
drf-yasg==1.21.4
